AP_Bootloader: fix differing sign comparisions

+++ Tools/AP_Bootloader/bl_protocol.cpp
@@ -657,7 +657,7 @@ bootloader(unsigned timeout)
                 goto cmd_bad;
             }
 
-            if (arg > sizeof(flash_buffer.c)) {
+            if ((size_t)arg > sizeof(flash_buffer.c)) {
                 goto cmd_bad;
             }
 
@@ -677,7 +677,7 @@ bootloader(unsigned timeout)
 
             // save the first words and don't program it until everything else is done
             if (address < sizeof(first_words)) {
-                uint8_t n = MIN(sizeof(first_words)-address, arg);
+                uint8_t n = MIN(sizeof(first_words) - address, (unsigned)arg);
                 memcpy(&first_words[address/4], &flash_buffer.w[0], n);
                 // replace first words with 1 bits we can overwrite later
                 memset(&flash_buffer.w[0], 0xFF, n);
