+++ Tools/AP_Bootloader/support.cpp
@@ -163,14 +163,14 @@ uint32_t get_mcu_desc(uint32_t max, uint8_t *revstr)
 
     mcu_des_t des = mcu_descriptions[STM32_UNKNOWN];
 
-    for (int i = 0; i < ARRAY_SIZE(mcu_descriptions); i++) {
+    for (unsigned i = 0; i < ARRAY_SIZE(mcu_descriptions); i++) {
         if (mcuid == mcu_descriptions[i].mcuid) {
             des = mcu_descriptions[i];
             break;
         }
     }
 
-    for (int i = 0; i < ARRAY_SIZE(silicon_revs); i++) {
+    for (unsigned i = 0; i < ARRAY_SIZE(silicon_revs); i++) {
         if (silicon_revs[i].revid == revid) {
             des.rev = silicon_revs[i].rev;
         }
@@ -202,7 +202,7 @@ bool check_limit_flash_1M(void)
     uint32_t idcode = (*(uint32_t *)DBGMCU_BASE);
     uint16_t revid = ((idcode & REVID_MASK) >> 16);
 
-    for (int i = 0; i < ARRAY_SIZE(silicon_revs); i++) {
+    for (unsigned i = 0; i < ARRAY_SIZE(silicon_revs); i++) {
         if (silicon_revs[i].revid == revid) {
             return silicon_revs[i].limit_flash_size_1M;
         }
@@ -295,7 +295,7 @@ void *memcpy(void *dest, const void *src, size_t n)
 {
     uint8_t *tdest = (uint8_t *)dest;
     uint8_t *tsrc = (uint8_t *)src;
-    for (int i=0; i<n; i++) {
+    for (size_t i=0; i<n; i++) {
         tdest[i] = tsrc[i];
     }
     return dest;
